A project on  Steganography with visualisation of the text within the cipher text
This project helps hide messages within data. The text it uses to hide
messages is the lorem ipsum.
This also includes the feature to add the encrypted key that can extract
the data within the file  in which the data is stored else the key can
be used separately to extract the data.